Big API Change: Splitted all configuration values to the places where they actually are required or where they belong.
This opens up the path to a "pure core" that does not now anything of other
applications.
I just changed the default component loading configuration values to
['inyoka.core.*',
'inyoka.*'], this is an aggressive default to ensure everything gets loaded.
I think that this is not the most proper way but it works for now. It can be
changed in the inyoka.ini anyway so one can change this to get a faster
startup-time (without the recursive package loading).

There's yet a lot of stuff to cleanup, move and to document. This will be the
context of following changes. Right now unittests pass properly (without any
change!) so that we can add new tests once we see any runtime changes (found
none yet)